Move mockSCFilePV
